https://doi.org/10.1093/bioinformatics/btz612 A SNP Effect Matrix (SEM) estimates the binding affinity of every possible mutation in a particular transcription factor (TF) binding motif This class contains three slots: the matrix, the baseline value, and a unique id

https://doi.org/10.1093/bioinformatics/btz612 A SNP Effect Matrix (SEM) estimates the binding affinity of every possible mutation in a particular transcription factor (TF) binding motif This class contains three slots: the matrix, the baseline value, and a unique id

show for SNPEffectMatrix

Usage

# S4 method for class 'SNPEffectMatrix'
show(object)

Arguments

object

SNPEffectMatrix object.

Slots

sem

The SEM itself as a data table Rows represent sequence position (variable length), columns represent effects due to each nucleotide base A, C, G, T (fixed length: 4)

baseline

A scrambled baseline, representing the binding score of randomly scrambled kmers of the same length. This is the binding cutoff for a TF

semId

basename of the sem file