Plot motif enrichment results

Plot motif enrichment results

Usage

plotSemEnrichment(
  e,
  lab = "semId",
  sigThreshold = 0.05,
  orThreshold = 1.25,
  sigCol = "dodgerblue2"
)

Arguments

e

SEMplR enrichment results data.table

lab

column in enrichment results to label significant points by. Default is "semId".

sigThreshold

adjusted pvalue threshold for coloring and labeling points. Default is 0.05

orThreshold

odds ratio threshold for coloring and labeling points. Default is 1.25

sigCol

color used to plot significant points

Value

a ggplot with sem scores for each nucleic acid per position