Calculate binding propensity for all SEM motifs and genomic positions provided

Calculate binding propensity for all SEM motifs and genomic positions provided

Usage

scoreBinding(x, sem, genome, nFlank = NULL, seqId = NULL)

Arguments

x

GRanges object or a vector of DNA sequences

sem

A SNPEffectMatrix or SNPEffectMatrixCollection object

genome

A BSgenome object for the genome build to use. ie. BSgenome.Hsapiens.UCSC.hg19::Hsapiens. Required if providing a GRanges object. Ignored if providing a vector of sequences.

nFlank

Number of flanking nucleotides to add to provided range. By default will add flank equal to the length of the longest motif. Ignored if providing a vector of sequences.

seqId

Column in GRanges object to use for unique id. By default, ids will be generated from the seqnames and ranges. Ignored if not providing a GRanges object.

Value

If a GRanges object is provided, return a SEMplScores object. If a list of sequences is provided, just return the scoring table

Examples

# load SEMs
data(sc)

# create a GRanges object
gr <- GenomicRanges::GRanges(
    seqnames = "chr12",
    ranges = 94136009
)

# calculate binding propensity
scoreBinding(gr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
#> An object of class SEMplScores
#> ranges(1): chr12:94136009
#> semData(13): transcription_factor, ensembl_id ... PWM_source, SEM_KEY
#> scores(223):
#>               seqId                SEM      score  scoreNorm index
#>              <char>             <char>      <num>      <num> <int>
#>   1: chr12:94136009 AP2B_HUMAN.SK-N-SH  -1.689754 -0.3068238    15
#>   2: chr12:94136009 ARNT_HUMAN.GM12878  -6.892799 -0.9693833    17
#>   3: chr12:94136009    ATF1_HUMAN.K562  -7.079925 -0.9420095    16
#>   4: chr12:94136009   ATF2_HUMAN.HepG2  -4.890126 -0.9098440    16
#>   5: chr12:94136009   ATF3_HUMAN.HepG2  -8.605675 -0.9885365    14
#>  ---                                                              
#> 219: chr12:94136009  ZBT7A_HUMAN.HepG2  -1.859506 -0.6349682    12
#> 220: chr12:94136009    ZFX_HUMAN.HepG2  -1.459472 -0.5682106    19
#> 221: chr12:94136009  ZN281_HUMAN.HepG2  -4.347612 -0.9355197     8
#> 222: chr12:94136009 ZNF18_HUMAN.HEK293  -5.410220 -0.9264060    15
#> 223: chr12:94136009   ZSCAN4_secondary -15.439087 -0.9998307    15
#>                   seq
#>                <char>
#>   1:       GCTTTGAGGC
#>   2:        TTTGAGGCA
#>   3:      CTTTGAGGCAT
#>   4:      CTTTGAGGCAT
#>   5:      GGCTTTGAGGC
#>  ---                 
#> 219:        AAGGCTTTG
#> 220:       TGAGGCATCT
#> 221:  GGAGAAGGCTTTGAG
#> 222:     GCTTTGAGGCAT
#> 223: GCTTTGAGGCATCTGC