Calculates binding enrichment for motif(s)

Perform a binomial test to determine if SNP Effect Matrices are bound more often than expected.

Usage

enrichSEMs(x, sem, background = NULL, seqs = NULL, nFlank = 0, genome = NULL)

Arguments

x

The scoring table produced by scoreBinding

sem

A SNPEffectMatrix or SNPEffectMatrixCollection object

background

A list of DNA sequences to use for background. The length of each sequence must match the length of sequences in x. By default, will scramble the provided sequences.

seqs

The sequences scored in scoreBinding

nFlank

Number of flanking nucleotides added to the sequences. Defaults to the length of the longest motif. If no flanks were added (ie, sequences were scored rather than a GRanges), use nFlank = 0.

genome

A BSGenome object. Requried if background isn't specified.

Value

a list of matrices

Examples

# load SEMs
data(sc)

# note that this is a small example for demonstration purposes
# in actual enrichment analyses sets of 100+ ranges are recommended

# create a GRanges object
gr <- GenomicRanges::GRanges(
    seqnames = "chr12",
    ranges = 94136009
)

# calculate binding propensity
sb <- scoreBinding(gr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

enrichSEMs(sb, sc)
#>                     SEM pvalue n_bound n_bound_bg  padj
#>                  <char>  <num>   <int>      <int> <num>
#>   1: AP2B_HUMAN.SK-N-SH      1       0          0     1
#>   2: ARNT_HUMAN.GM12878      1       0          0     1
#>   3:    ATF1_HUMAN.K562      1       0          0     1
#>   4:   ATF2_HUMAN.HepG2      1       0          0     1
#>   5:   ATF3_HUMAN.HepG2      1       0          0     1
#>  ---                                                   
#> 219:  ZBT7A_HUMAN.HepG2      1       0          0     1
#> 220:    ZFX_HUMAN.HepG2      1       0          0     1
#> 221:  ZN281_HUMAN.HepG2      1       0          0     1
#> 222: ZNF18_HUMAN.HEK293      1       0          0     1
#> 223:   ZSCAN4_secondary      1       0          0     1