Calculate risk/non-risk binding propensity for all SEM motifs and variants provided

Calculate risk/non-risk binding propensity for all SEM motifs and variants provided

Usage

scoreVariants(x, sem, genome, refCol = NULL, altCol = NULL, varId = NULL)

Arguments

x

VRanges object

sem

a list of SNPEffectMatrix objects

genome

A BSgenome object for the genome build to use. ie. BSgenome.Hsapiens.UCSC.hg19::Hsapiens

refCol

If providing a GRanges, the meta data column name with the reference (ref) allele. Ignored if providing a VRanges object.

altCol

If providing a GRanges, the meta data column name with the alternative (alt) allele. Ignored if providing a VRanges object.

varId

A column name in the meta data of x to use as a unique id.

Value

a SEMplScores object

Examples

library(VariantAnnotation)

# load default SEMs
data(sc)

# create a VRanges object
x <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
scoreVariants(x, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
#> An object of class SEMplScores
#> ranges(1): chr12:94136009:G>C
#> semData(13): transcription_factor, ensembl_id ... PWM_source, SEM_KEY
#> scores(223):
#>                   varId              semId           refSeq           altSeq
#>                  <char>             <char>           <char>           <char>
#>   1: chr12:94136009:G>C AP2B_HUMAN.SK-N-SH       GCTTTGAGGC       GCTTTCAGGC
#>   2: chr12:94136009:G>C ARNT_HUMAN.GM12878        TTTGAGGCA        TTTCAGGCA
#>   3: chr12:94136009:G>C    ATF1_HUMAN.K562      CTTTGAGGCAT      TTCAGGCATCT
#>   4: chr12:94136009:G>C   ATF2_HUMAN.HepG2      CTTTGAGGCAT      CTTTCAGGCAT
#>   5: chr12:94136009:G>C   ATF3_HUMAN.HepG2      GGCTTTGAGGC      TTTCAGGCATC
#>  ---                                                                        
#> 219: chr12:94136009:G>C  ZBT7A_HUMAN.HepG2        AAGGCTTTG        AAGGCTTTC
#> 220: chr12:94136009:G>C    ZFX_HUMAN.HepG2       TGAGGCATCT       TCAGGCATCT
#> 221: chr12:94136009:G>C  ZN281_HUMAN.HepG2  GGAGAAGGCTTTGAG  AAGGAGAAGGCTTTC
#> 222: chr12:94136009:G>C ZNF18_HUMAN.HEK293     GCTTTGAGGCAT     GGCTTTCAGGCA
#> 223: chr12:94136009:G>C   ZSCAN4_secondary GCTTTGAGGCATCTGC GCTTTCAGGCATCTGC
#>        refScore   altScore    refNorm    altNorm refVarIndex altVarIndex
#>           <num>      <num>      <num>      <num>       <int>       <int>
#>   1:  -1.689754  -1.994456 -0.3068238 -0.4387980          15          15
#>   2:  -6.892799  -4.947151 -0.9693833 -0.8820613          17          17
#>   3:  -7.079925  -7.978780 -0.9420095 -0.9688990          16          18
#>   4:  -4.890126  -4.983040 -0.9098440 -0.9154673          16          16
#>   5:  -8.605675  -8.098832 -0.9885365 -0.9837111          14          17
#>  ---                                                                    
#> 219:  -1.859506  -1.607344 -0.6349682 -0.5652506          12          12
#> 220:  -1.459472  -1.478959 -0.5682106 -0.5740037          19          19
#> 221:  -4.347612  -4.740351 -0.9355197 -0.9508865           8           6
#> 222:  -5.410220  -7.107575 -0.9264060 -0.9773072          15          14
#> 223: -15.439087 -13.002975 -0.9998307 -0.9990837          15          15